==> notebooks/Notebook0.Finding Intersection between Muck Rack and GDELT (and MBFC).ipynb <==
# Finding Intersection between Muck Rack and GDELT

"What counts as a news source?" is an important question. Muck Rack and GDELT have their own definition of what can be considered as a news source. An intersection between a set of manually curated news source selection from Muck Rack, and a set of news sources recognized by the global database GDELT will provide us with news sources that are recognized by at least two institutions.

Table of content

1. [Read news sources from GDELT](#sec1)
    * [Does GDELT have domains and subdomains OR just domains?](#sec1a) Gdelt has subdomains!
    
2. [Read news sources from Muck Rack](#sec2)
    * [New Muck Rack](#sec2a)
    * [Old Muck Rack](#sec2b)
    * [Union between old and new Muck Rack](#sec2c)

3. [Intersection between Muck Rack and GDELT](#sec3)
4. [Intersection between Muck Rack, GDELT, and MBFC](#sec4)

import json
import csv

from collections import defaultdict, Counter

import tldextract

<a id="sec1"></a>
## Read news sources from GDELT

years = [2016, 2017, 2018, 2019, 2020, 2021]

gdelt_ns = set()

for year in years:
    with open (f"D:\\Wellesley\\F21\\thesis\\data\\gdelt\\gdelt_{year}.csv", "r", encoding="utf-8") as gi:
        reader = csv.DictReader(gi)
        for row in reader:
            gdelt_ns.add(row['SourceCommonName'])

gdelt_ns = list(gdelt_ns)
len(gdelt_ns)

gdelt_ns[:10]

<a id="sec1a"></a>
### Does GDELT have domains and subdomains OR just domains? YES, with subdomains!

with_subdomains = []

for gn in gdelt_ns:
    extracted = tldextract.extract(gn)
    subdomain, domain, suffix = extracted
    if subdomain:
        with_subdomains.append(gn)
        
print(f"there are {len(with_subdomains)} news sources with subdomains.")

### Strip www. from gdelt news sources

gdelt_ns = list(set([k.replace("www.","") for k in gdelt_ns]))

len(gdelt_ns)

<a id="sec2"></a>
## Read news sources from Muck Rack

<a id="sec2a"></a>
### First, new Muck Rack

(scraped in January 2022)

with open("D:\\Wellesley\\F21\\thesis\\data\\muckrack\\all_new_muckrack.json", "r", encoding="utf-8") as mrinfile:
    new_muckrack = json.load(mrinfile)

new_muckrack_ns = list(new_muckrack.keys())
new_muckrack_ns[:10]

len(set(new_muckrack_ns))

#### Let's clean it

clean_new_muckrack = set()

for nsm in new_muckrack_ns:
    extracted = tldextract.extract(nsm)
    subdomain, domain, suffix = extracted
    # add both versions of domain.suffix and subdomain.domain.suffix
    full = ""
    # with subdomain
    if len(subdomain) > 0:
        #print(f"{subdomain}.{domain}.{suffix}")
        full = f"{subdomain}.{domain}.{suffix}"
        if len(full) > 0:
            clean_new_muckrack.add(full.replace("www.","").strip('/'))
    # without subdomain
    full = f"{domain}.{suffix}"
    if len(full) > 0:
        clean_new_muckrack.add(full.replace("www.","").strip('/'))
len(clean_new_muckrack)

list(clean_new_muckrack)[:10]

<a id="sec2b"></a>

### Old Muck Rack

with open("D:\\Wellesley\\F21\\thesis\\data\\muckrack\\all_old_muckrack.json", "r", encoding="utf-8") as mrinfile1:
    old_muckrack = json.load(mrinfile1)

old_muckrack_ns = list(old_muckrack.keys())
old_muckrack_ns[:10]

#### Let's clean it

len(set(old_muckrack_ns))

clean_old_muckrack = set()

for lsm in old_muckrack_ns:
    extracted = tldextract.extract(lsm)
    subdomain, domain, suffix = extracted
    # add both versions of domain.suffix and subdomain.domain.suffix
    full = ""
    # with subdomain
    if len(subdomain) > 0:
        #print(f"{subdomain}.{domain}.{suffix}")
        full = f"{subdomain}.{domain}.{suffix}"
        if len(full) > 0:
            clean_old_muckrack.add(full.replace("www.","").strip('/'))
    # without subdomain
    full = f"{domain}.{suffix}"
    if len(full) > 0:
        clean_old_muckrack.add(full.replace("www.","").strip('/'))
len(clean_old_muckrack)

list(clean_old_muckrack)[:10]

<a id="sec2c"></a>
### Union between Old and New Muck Rack

len(clean_old_muckrack & clean_new_muckrack)

import difflib

difflib.get_close_matches('995magicfm.com', list(old_muckrack.keys()))

difflib.get_close_matches('http://995magicfm.com', list(new_muckrack.keys()))

muckrack_union = clean_old_muckrack | clean_new_muckrack

muckrack_union.remove("")

len(muckrack_union)

# with open ("D:\\Wellesley\\F21\\thesis\\data\\muckrack\\muckrack_union.json", "w", encoding="utf-8") as unionfile:
#     json.dump(list(muckrack_union), unionfile, indent=4)

<a id="sec3"></a>
## Intersection between Muck Rack and GDELT

gm_intersection = set(gdelt_ns) & set(muckrack_union)
len(gm_intersection)

# with open ("D:\\Wellesley\\F21\\thesis\\data\\gm_intersection.json", "w", encoding="utf-8") as outfile:
#     json.dump(list(gm_intersection), outfile, indent=4)

len(gdelt_ns)

len(set(gdelt_ns) - gm_intersection)

len(gm_intersection)/len(gdelt_ns)

len(muckrack_union)

len(set(muckrack_union) - gm_intersection)

len(gm_intersection)/len(muckrack_union)

<a id="sec4"></a>
## Intersection with MBFC

categories = ["center", "conspiracy", "fake-news", "left", "leftcenter", "pro-science", "pseudoscience-dictionary", "right", "right-center", "satire"]

mbfc_all = set()
mbfc_divided = defaultdict(set)

for c in categories:
    with open(f"D:\\Wellesley\\F21\\thesis\\data\\mbfc\\mbfc_{c}.json", "r", encoding="utf-8") as infile:
        file = json.load(infile)
        for ns in file:
            source_name, source_link, page_link = ns
            mbfc_all.add(source_link) # I will clean later
            mbfc_divided[c].add(source_link)

len(mbfc_all)

list(mbfc_all)[:10]

### Cleaning `mbfc_all`

mbfc_summary = set()

for ns in mbfc_all:
    extracted = tldextract.extract(ns)
    subdomain, domain, suffix = extracted
    # add both versions of domain.suffix and subdomain.domain.suffix
    full = ""
    # with subdomain
    if len(subdomain) > 0:
        #print(f"{subdomain}.{domain}.{suffix}")
        full = f"{subdomain}.{domain}.{suffix}"
        if len(full) > 0:
            mbfc_summary.add(full.replace("www.","").strip('/'))
    # without subdomain
    full = f"{domain}.{suffix}"
    if len(full) > 0:
        mbfc_summary.add(full.replace("www.","").strip('/'))
len(mbfc_summary)
    

list(mbfc_summary)[:10]

with open("D:\\Wellesley\\F21\\thesis\\data\\mbfc\\mbfc_summary.json", "w", encoding="utf-8") as outfile1:
    json.dump(list(mbfc_summary), outfile1, indent=4)

#### Intersection with gdelt and muck rack

len(mbfc_summary & set(gm_intersection))

gmm_intersection = list(mbfc_summary & set(gm_intersection))

with open("D:\\Wellesley\\F21\\thesis\\data\\gmm_intersection.json", "w", encoding="utf-8") as outfile2:
    json.dump(gmm_intersection, outfile2)
