==> notebooks/Notebook1a.Get all relevant submissions.ipynb <==
# Get all relevant submissions

Author: Junita Sirait

I am tired of parsing the whole 6 months of data each time I need additional informations from the JSON files. Here I will master the art of `ndjson` dump and load, so I can just process these files in the future.

Table of contents:
1. [Reading the original massive files](#sub1)

# !pip install ndjson

import ndjson

import json
import pandas as pd
import zstandard as zstd
import io

from collections import defaultdict, Counter
from urllib.parse import urlparse
import re
import datetime, time
import tldextract

## Reading the original massive files

with open("D:\\Wellesley\\F21\\thesis\\data\\gmm_intersection.json", "r") as infile:
    gmm_news_sources = json.load(infile)

dctx = zstd.ZstdDecompressor(max_window_size=2147483648)

def findURLs(phrase):
    regex = re.compile('((https?):((//)|(\\\\))+([\w\d:#@%/;$()~_?\+-=\\\.&](#!)?)*)')
    url = re.findall(regex, phrase)     
    return [x[0] for x in url]

def get_hostname(url, uri_type='both'):
    """Get the host name from the url"""
    hostnames = set()
    extracted = tldextract.extract(url)
    subdomain, domain, suffix = extracted
    # add both versions of domain.suffix and subdomain.domain.suffix
    full = ""
    # with subdomain
    if len(subdomain) > 0 and len(suffix) > 0:
        #print(f"{subdomain}.{domain}.{suffix}")
        full = f"{subdomain}.{domain}.{suffix}"
        if len(full) > 0:
            hostnames.add(full[4:].strip('/')) if full.startswith("www.") else hostnames.add(full.strip('/'))
    # without subdomain
    full = f"{domain}.{suffix}"
    if len(full) > 0 and len(suffix) > 0:
        hostnames.add(full[4:].strip('/')) if full.startswith("www.") else hostnames.add(full.strip('/'))
    return hostnames

zst_files = ["RS_2021-02.zst", "RS_2021-03.zst", "RS_2021-04.zst", "RS_2021-05.zst", "RS_2021-06.zst", "RS_2021-01.zst"]
zst_filepath = "E:/thesis_data/"

subreddits_total_activity = defaultdict(dict)

print("start time:", datetime.datetime.now())


for zst_file in zst_files:
    counter = 0
    added = 0
    month = zst_file[8:10]
    with open ("E:/relevant_posts_{}.ndjson".format(month), "a", encoding="utf-8") as ndjfile:
        writer = ndjson.writer(ndjfile, ensure_ascii=False)
        print("***** Start processing for {} *****".format(zst_file))
        with open(zst_filepath+zst_file, 'rb') as ifh:
            with dctx.stream_reader(ifh, read_size=2) as reader:
                text_stream = io.TextIOWrapper(reader, encoding='utf-8')
                for d in text_stream:
                    line = json.loads(d)
                    URLs = findURLs(line['url']) + findURLs(line['selftext'])                
                    hostnames = [get_hostname(url) for url in URLs]
                    URLs = [element for sub in hostnames for element in sub]
                    subreddit = line["subreddit"]
                    if month in subreddits_total_activity[subreddit]:
                        subreddits_total_activity[subreddit][month] += 1
                    else:
                        subreddits_total_activity[subreddit][month] = 1
                    # print(URLs)
                    for url in URLs:
                        if url in gmm_news_sources:
                            writer.writerow(line)
                            added += 1
                            break # write this post only once
                    url_of_our_ns = False
                    counter += 1
                    if counter%500000 == 0: 
                        print("processed {}; added {}; \t by {}".format(counter, added, str(datetime.datetime.now())[11:19]))
        
        print("****************************************** Summary ******************************************")
        print(f"There are {counter} total posts in {month}, and {added} of them have urls to our news sources.")
        
        print("----------------------------------------------------------------------------------------")
        print("-------------------------------- Done processing for {} --------------------------------".format(zst_file))
        print("----------------------------------------------------------------------------------------")
                
print("finish time:", datetime.datetime.now())

len(subreddits_total_activity)

len([s for s in subreddits_total_activity if len(subreddits_total_activity[s])==6])

subreddits_total_activity["twilight"]

with open("subreddits_total_activity.json", "w", encoding="utf-8") as af:
    json.dump(subreddits_total_activity, af)

```
Month            |    with news  |      total    |    % url with news    |
-------------------------------------------------------------------------|
January          |    1137576    |    32704571   |         3.48%         |
February         |    1026958    |    31147947   |         3.29%         |
March            |    1155554    |    33006103   |         3.50%         |
April            |    1090699    |    31616206   |         3.45%         |
May              |    1019503    |    36310673   |         2.81%         |
June             |     811758    |    34118481   |         2.38%         |
```

811758/34118481

"wikipedia.org" in gmm_news_sources

counter
