==> notebooks/Notebook1b.Subreddit Activity.ipynb <==
# Subreddit Activity

Author: Junita Sirait



import ndjson

import json
import pandas as pd
import zstandard as zstd
import io

from collections import defaultdict, Counter
from urllib.parse import urlparse
import re
import datetime, time
import tldextract

with open("subreddits_total_activity1.json", "r", encoding="utf-8") as af:
    subreddits_activity = json.load(af)

sorted([s for s in subreddits_activity.keys()], key=lambda a:sum(subreddits_activity[a].values()), reverse=True)

subreddits_activity["u_toronto_news"]

print(len([k for k in subreddits_activity if len(subreddits_activity[k])!=6]))

print(len([k for k in subreddits_activity if len(subreddits_activity[k])==6])) # active in all six months

sorted([s for s in subreddits_activity.keys()], key=lambda a:subreddits_activity[a]['01'] if '01' in subreddits_activity[a] else 0, reverse=True)[:20]

sorted([s for s in subreddits_activity.keys()], key=lambda a:subreddits_activity[a]['02'] if '02' in subreddits_activity[a] else 0, reverse=True)[:20]

sorted([s for s in subreddits_activity.keys()], key=lambda a:subreddits_activity[a]['03'] if '03' in subreddits_activity[a] else 0, reverse=True)[:20]

sorted([s for s in subreddits_activity.keys()], key=lambda a:subreddits_activity[a]['04'] if '04' in subreddits_activity[a] else 0, reverse=True)[:20]

sorted([s for s in subreddits_activity.keys()], key=lambda a:subreddits_activity[a]['05'] if '05' in subreddits_activity[a] else 0, reverse=True)[:20]

sorted([s for s in subreddits_activity.keys()], key=lambda a:subreddits_activity[a]['06'] if '06' in subreddits_activity[a] else 0, reverse=True)[:20]
