==> notebooks/3 - Subreddits for 6k news sources.ipynb <==
# Preliminary look at subreddits for 6k news sources

!pip install zstandard

import json
import pandas as pd
import zstandard as zstd
import io
from collections import defaultdict, Counter
from urllib.parse import urlparse
import re
import datetime, time

print(datetime.datetime.now())
print(str(datetime.datetime.now())[11:19])

with open("gmw_intersection.json", "r") as infile:
    news_sources = json.load(infile)

len(news_sources)

news_sources

## Open reddit data from April 2021

Reference: https://arxiv.org/pdf/2001.08435.pdf

Example of an entry of data:

```
{
    'all_awardings': [], 
    'allow_live_comments': False, 
    'archived': False, 
    'author': 'elanglohablante9805', 
    'author_created_utc': 1609519842, 
    'author_flair_background_color': '#ffb000', 
    'author_flair_css_class': None, 
    'author_flair_richtext': [], 
    'author_flair_template_id': '4f908eaa-9664-11ea-a567-0ed46a42aec3', 
    'author_flair_text': 'Historiador 📜 | 80-Day Streak 🔥', 
    'author_flair_text_color': 'dark', 
    'author_flair_type': 'text', 
    'author_fullname': 't2_9lr431i4', 
    'author_patreon_flair': False, 
    'author_premium': False, 
    'can_gild': True, 
    'category': None, 
    'content_categories': None, 
    'contest_mode': False, 
    'created_utc': 1617235201, 
    'discussion_type': None, 
    'distinguished': None, 
    'domain': 'self.WriteStreakES', 
    'edited': False, 
    'gilded': 0, 
    'gildings': {}, 
    'hidden': False, 
    'hide_score': False, 
    'id': 'mhj2hj', 
    'is_created_from_ads_ui': False, 
    'is_crosspostable': True, 
    'is_meta': False, 
    'is_original_content': False, 
    'is_reddit_media_domain': False, 
    'is_robot_indexable': True, 
    'is_self': True, 
    'is_video': False, 
    'link_flair_background_color': '', 
    'link_flair_css_class': None, 
    'link_flair_richtext': [], 
    'link_flair_text': None, 
    'link_flair_text_color': 'dark', 
    'link_flair_type': 'text', 
    'locked': False,
    'media': None, 
    'media_embed': {}, 
    'media_only': False, 
    'name': 't3_mhj2hj', 
    'no_follow': True, 
    'num_comments': 2, 
    'num_crossposts': 0, 
    'over_18': False, 
    'parent_whitelist_status': None, 
    'permalink': '/r/WriteStreakES/comments/mhj2hj/streak_90_ha_llegado_la_primavera/', 
    'pinned': False, 
    'pwls': None, 
    'quarantine': False, 
    'removed_by_category': None, 
    'retrieved_utc': 1623447663, 
    'score': 1, 
    'secure_media': None, 
    'secure_media_embed': {}, 
    'selftext': 'Los pájaros están cantando, las hierbas verdes están brotando, y tengo alergias.  Esto es la temporada de las alergias.  Estornudo cada mañana cuando me despierto, y otra vez si voy afuera.  Necesito tomar medicina cada día, pero no funciona tan bien. \n\nPor fuera, las lomas son bonitas porque son verdes y los robles tienen hojas nuevas.  Por el fin de semana,  hago caminatas pero cuando regreso a casa, necesito ducharme para remover el polen.\n\nCuando me jubile, voy a viajar al desierto cada año por toda la primavera.  No me gustaría quedarme aquí.', 
    'send_replies': True, 
    'spoiler': False, 
    'stickied': False, 
    'subreddit': 'WriteStreakES', 
    'subreddit_id': 't5_2eamt5', 
    'subreddit_subscribers': 2205, 
    'subreddit_type': 'public', 
    'suggested_sort': None, 
    'thumbnail': 'self', 
    'thumbnail_height': None, 
    'thumbnail_width': None, 
    'title': 'Streak 90: Ha llegado la primavera', 
    'top_awarded_type': None, 
    'total_awards_received': 0, 
    'treatment_tags': [], 
    'upvote_ratio': 1.0, 
    'url': 'https://www.reddit.com/r/WriteStreakES/comments/mhj2hj/streak_90_ha_llegado_la_primavera/', 
    'whitelist_status': None, 'wls': None}

```

dctx = zstd.ZstdDecompressor(max_window_size=2147483648)

ns_subreddit = defaultdict(Counter) # counting how many time a news source appears in each subreddit
subreddit_ns = defaultdict(Counter)
subreddit_srid = dict()

posts_with_urls = list()
posts_with_urls

def findURLs(phrase):
    regex = re.compile('((https?):((//)|(\\\\))+([\w\d:#@%/;$()~_?\+-=\\\.&](#!)?)*)')
    url = re.findall(regex, phrase)     
    return [x[0] for x in url]

# try out
findURLs("does this find https://lol or nytimes.com/2021/10/19/us/politics/trump-border.html")

def get_hostname(url, uri_type='both'):
    """Get the host name from the url"""
    domain = re.compile(r"(https?://)?(www\.)?")
    return domain.sub('', url).strip().strip('/').split('/')[0]

# function try out
print(get_hostname("https://www.nytimes.com"))
print(get_hostname("http://www.nytimes.com/add"))
print(get_hostname("www.nytimes.com/additional"))

"realtor.com" in news_sources

# zst_files = ["RS_2021-02.zst", "RS_2021-03.zst", "RS_2021-05.zst", "RS_2021-06.zst"]
zst_files = ["RS_2021-05.zst", "RS_2021-06.zst"]
zst_filepath = "D:/thesis_data/" # D for ThinkPad

print("start time:", datetime.datetime.now())

counter = 0
for zst_file in zst_files:
    print("***** Start processing for {} *****".format(zst_file))
    with open(zst_filepath+zst_file, 'rb') as ifh: #, open("stream_output.json", 'wb') as ofh:
        with dctx.stream_reader(ifh, read_size=2) as reader:
            text_stream = io.TextIOWrapper(reader, encoding='utf-8')
            url_of_our_ns = False 
            # ^this is to keep track if this post has ns url that we care about.
            # if so, add to posts_with_urls once.
            for d in text_stream:
                line = json.loads(d)
                subreddit, subreddit_id = line['subreddit'], line['subreddit_id']
                if subreddit not in subreddit_srid:
                    subreddit_srid[subreddit] = subreddit_id
                URLs = findURLs(line['url']) + findURLs(line['selftext'])
                URLs = [get_hostname(url) for url in URLs]
                # print("URLs:", URLs)
                # if len(URLs) > 10: print(line['selftext'])
                for url in URLs:
                    if url in news_sources:
                        if not url_of_our_ns:
                            posts_with_urls.append(line)
                            url_of_our_ns = True
                        ns_subreddit[url][subreddit] += 1
                        subreddit_ns[subreddit][url] += 1
                url_of_our_ns = False
                counter += 1
                if counter%500000 == 0: 
                    print("processed {} by {}".format(counter, str(datetime.datetime.now())[11:19]))
                
    
    print("-------------------------------- Done reading, will write files now --------------------------------")
    
    # write into files separated by months
    with open("ns_subreddit_{}.json".format(zst_file[3:10]), "w", encoding="utf-8") as outfile:
        json.dump(ns_subreddit, outfile)
        
    with open("subreddit_ns_{}.json".format(zst_file[3:10]), "w", encoding = "utf-8") as outfile1:
        json.dump(subreddit_ns, outfile1)
        
    with open ("subreddit_srid_{}.json".format(zst_file[3:10]), "w", encoding = "utf-8") as infile_srid:
        json.dump(subreddit_srid, infile_srid)
        
    with open("D:/thesis_data/posts_with_urls_{}.json".format(zst_file[3:10]), "w", encoding = "utf-8") as outfile2:
        json.dump(posts_with_urls, outfile2)
        
    ns_subreddit = defaultdict(Counter) # counting how many time a news source appears in each subreddit
    subreddit_ns = defaultdict(Counter)
    # subreddit_srid = dict()
    posts_with_urls = list()
    counter = 0
        
    print("----------------------------------------------------------------------------------------")
    print("-------------------------------- Done processing for {} --------------------------------".format(zst_file))
    print("----------------------------------------------------------------------------------------")
                
print("finish time:", datetime.datetime.now())

# with open ("subreddit_srid_{}.json".format(zst_file[3:10]), "w", encoding = "utf-8") as infile_srid:
#     json.dump(subreddit_srid, infile_srid)
        
# with open("D:/thesis_data/posts_with_urls_{}.json".format(zst_file[3:10]), "w", encoding = "utf-8") as outfile2:
#     json.dump(posts_with_urls, outfile2)

zst_file[3:10]

counter

Number of posts read from 
1. January 2021:
2. February 2021: 31,161,912
3. March 2021: 33,0061,03
4. April 2021:
5. May 2021: >36M
6. June 2021: >34M

with open ("subreddit_srid_{}.json".format("2021-04"), "w", encoding = "utf-8") as infile_srid:
    json.dump(subreddit_srid, infile_srid)

len(posts_with_urls)

len(ns_subreddit)

3686 news sources

len(subreddit_ns)

20755 subreddits

len(posts_with_urls)

with open("ns_subreddit.json", "w", encoding="utf-8") as outfile:
    json.dump(ns_subreddit, outfile)

with open("subreddit_ns.json", "w", encoding = "utf-8") as outfile1:
    json.dump(subreddit_ns, outfile1)

with open("posts_with_urls", "w", encoding = "utf-8") as outfile2:
    json.dump(posts_with_urls, outfile2)

counter

with open("counter_april21.json", "w", encoding = "utf-8") as counterfile:
    json.dump(counter, counterfile)

**Getting subreddit names and ids**

subreddit_id = defaultdict(set)
id_subreddit = defaultdict(set)

print("start time:", datetime.datetime.now())

counter = 0
with open("D://Wellesley/F21/thesis_zst_data/RS_2021-04.zst", 'rb') as ifh: #, open("stream_output.json", 'wb') as ofh:
    with dctx.stream_reader(ifh, read_size=2) as reader:
        text_stream = io.TextIOWrapper(reader, encoding='utf-8')
        for d in text_stream:
            line = json.loads(d)
            sr, sr_id = line['subreddit'], line['subreddit_id']
            subreddit_id[sr].add(sr_id)
            id_subreddit[sr_id].add(sr)
#             URLs = findURLs(line['url']) + findURLs(line['selftext'])
#             URLs = [get_hostname(url) for url in URLs]
#             # print("URLs:", URLs)
#             # if len(URLs) > 10: print(line['selftext'])
#             for url in URLs:
#                 if url in news_sources:
#                     posts_with_urls.append(line)
#                     ns_subreddit[url][subreddit] += 1
#                     subreddit_ns[subreddit][url] += 1
            counter += 1
            if counter%500000 == 0: print(f"at {counter}")
                
print("finish time:", datetime.datetime.now())

len(subreddit_id)

len(id_subreddit)

for s in subreddit_id:
    if len(subreddit_id[s]) != 1:
        print(f"{s} is invalid, length {len(subreddit_id[s])}")

for s in id_subreddit:
    if len(id_subreddit[s]) != 1:
        print(f"{s} is invalid, length {len(id_subreddit[s])}")

Let's make all values to be strings.

for s in subreddit_id:
    subreddit_id[s] = list(subreddit_id[s])[0]

for i in id_subreddit:
    id_subreddit[i] = list(id_subreddit[i])[0]

There are the same number of `id`s and `subreddit`s. Good!

with open("subreddit_id.json", "w", encoding = "utf-8") as outfile_si:
    json.dump(subreddit_id, outfile_si)

with open("id_subreddit.json", "w", encoding = "utf-8") as outfile_is:
    json.dump(id_subreddit, outfile_is)
