==> notebooks/6 - Getting subreddits details.ipynb <==
# Getting Details of Subreddits

I will be investigating the details of subreddits provided by PushShift which is accessible here: https://files.pushshift.io/reddit/subreddits/

!pip install ijson

import json
import csv
import pandas as pd

import zstandard as zstd
import io

from collections import Counter, defaultdict

import ijson # to stream large json files

import datetime

dctx = zstd.ZstdDecompressor(max_window_size=2147483648)

dict_list = []

counter = 0
# with open("D://Wellesley/F21/thesis_zst_data/reddit_subreddits.ndjson.zst", 'rb') as ifh: #, open("stream_output.json", 'wb') as ofh:
with open("D://thesis_data/reddit_subreddits.ndjson.zst", 'rb') as ifh: #, open("stream_output.json", 'wb') as ofh:
    with dctx.stream_reader(ifh, read_size=2) as reader:
        text_stream = io.TextIOWrapper(reader, encoding='utf-8')
        for line in text_stream:
            if counter >= 1000:
                break
            d = json.loads(line)
            dict_list.append(d)
            counter += 1

ifh.close()

dict_list[0].keys()

Interesting keys:
1. description
2. public_description
3. public_traffic

print([dict_list[k]['subscribers'] for k in range(len(dict_list))])

dict_list[1]['description']

dict_list[1]['public_description']

Interesting how `description` is not exactly the same as `public_description`. Should I use both? Perhaps just `description`, because people are more inclined to post on a subreddit if they are a member of that subreddit instead of just a random general public?

[dict_list[idx]['title'] for idx in range(len(dict_list))]

[dict_list[idx]['id'] for idx in range(len(dict_list))]

It seems that I need subreddit id's afterall. Good thing I have this info. For each of the subreddits that we are interested in, let's get their id's and then descriptions.

with open("subreddit_ns.json", "r", encoding="utf-8") as infile1:
    subreddit_ns = json.load(infile1)
    subreddits = subreddit_ns.keys()
infile1.close()

len(subreddits)

Let's get their id's!

spec_sr_id = dict()
with open("subreddit_id.json", "r") as infile2:
    subreddit_id = json.load(infile2)
    for s in subreddits:
        spec_sr_id[s] = subreddit_id[s]

len(spec_sr_id)

spec_id_sr = {k:v for v,k in spec_sr_id.items()}

`spec_sr_id` contains the matching {`subreddit`: `it's id`}, and `spec_id_sr` is its inverse.

Now let's create the matching {`subreddit`: `it's description`} and {`subreddit_id`: `it's description`}

ids = list(spec_id_sr.keys())

len(ids)

ids[1034:1041] # what's with the t5_?

for i in ids:
    if i[:3] != 't5_':
        print(i)

Noting is printed from the above cell. So, in `ids`, all `subreddit_id` starts with `t5_`. Let's trim this then.

ids = [i[3:] for i in ids]

**Finding subreddit descriptions**

counter = 0
desc_found = 0
subreddit_description = dict()
id_description = dict()

print("start time: {}".format(datetime.datetime.now()))

with open("D://Wellesley/F21/thesis_zst_data/reddit_subreddits.ndjson.zst", 'rb') as ifh: #, open("stream_output.json", 'wb') as ofh:
    with dctx.stream_reader(ifh, read_size=2) as reader:
        text_stream = io.TextIOWrapper(reader, encoding='utf-8')
        for line in text_stream:
            # if counter >= 10000:
            #     break
            if counter % 50000 == 0:
                print("counter: {}".format(counter))
            d = json.loads(line)
            srdtid = d["id"]
            if srdtid in ids:
                # in spec_id_sr, id starts with t5_
                for_spec = "t5_" + srdtid
                subreddit_description[spec_id_sr[for_spec]] = d["description"]
                # id_description[for_spec] = d["description"]
                desc_found += 1
            counter += 1
            # let's break if descriptions for all id's have been found
            if desc_found == len(ids):
                print("all found after {} look ups. Done!".format(counter))
                break
            
print("finish time: {}".format(datetime.datetime.now()))

**Finding subreddit subscribers**

counter = 0
subscribers_found = 0
subreddit_subscribers = dict()
id_subscribers = dict()

print("start time: {}".format(datetime.datetime.now()))

# with open("D://Wellesley/F21/thesis_zst_data/reddit_subreddits.ndjson.zst", 'rb') as ifh: #, open("stream_output.json", 'wb') as ofh:
with open("D://thesis_data/reddit_subreddits.ndjson.zst", 'rb') as ifh: #, open("stream_output.json", 'wb') as ofh:
    with dctx.stream_reader(ifh, read_size=2) as reader:
        text_stream = io.TextIOWrapper(reader, encoding='utf-8')
        for line in text_stream:
            # if counter >= 10000:
            #     break
            if counter % 50000 == 0:
                print("counter: {}".format(counter))
            d = json.loads(line)
            srdtid = d["id"]
            if srdtid in ids:
                # in spec_id_sr, id starts with t5_
                for_spec = "t5_" + srdtid
                subreddit_subscribers[spec_id_sr[for_spec]] = d["subscribers"]
                id_subscribers[for_spec] = d["subscribers"]
                subscribers_found += 1
            counter += 1
            # let's break if descriptions for all id's have been found
            if subscribers_found == len(ids):
                print("all found after {} look ups. Done!".format(counter))
                break
            
print("finish time: {}".format(datetime.datetime.now()))

print(f"{subscribers_found} found out of {len(ids)} ({subscribers_found*100/len(ids)}). Counter: {counter}.")

print(f"{desc_found} found out of {len(ids)} ({desc_found*100/len(ids)}). Counter: {counter}.")

It is interesting that PushShift's `subreddit` folder only contains 70% of the subreddits mentioned in their `submission` folder. `subreddit` has 914k subreddits.

Perhaps 30% has no descriptions?

# example description
subreddit_description["prisons"]

spec_sr_id["prisons"]

id_description = {spec_sr_id[s]: subreddit_description[s] for s in subreddit_description}

id_description['t5_2sbiw']

Let's save these into files

with open("subreddit_description.json", "w", encoding = "utf-8") as outfile_srd:
    json.dump(subreddit_description, outfile_srd)

with open("id_description.json", "w", encoding = "utf-8") as outfile_id:
    json.dump(id_description, outfile_id)

with open("subreddit_subscribers.json", "w", encoding = "utf-8") as outfile_subc:
    json.dump(subreddit_subscribers, outfile_subc)

len(id_subscribers)

with open("id_subscribers.json", "w", encoding = "utf-8") as outfile_subc_id:
    json.dump(id_subscribers, outfile_subc_id)
