==> notebooks/Notebook12.Credible v. Not Credible based on MBFC.ipynb <==
# Credible v. Not Credible Based on MBFC

I will construct a file mapping each source to 1 if credible, 0 if not credible based on MBFC

import os

import json
import csv

from collections import defaultdict

# open credible news sources
pd1 = os.path.split(os.getcwd())[0]
mbfc_path = os.path.join(pd1, "data/mbfc")

c = json.load(open(os.path.join(mbfc_path, "mbfc_center.json")))
ps = json.load(open(os.path.join(mbfc_path, "mbfc_pro-science.json")))

centers = [k[1] for k in c]
len(centers)

proscience = [k[1] for k in ps]
len(proscience)

cr = list(set(centers) | set(proscience))
len(cr)

ns_fixed_ = json.load(open("ns_fixed_.json", "r", encoding="utf-8"))
len(ns_fixed_)

credible = list(set(cr) & set(ns_fixed_))
len(credible)

# open nonreliable news sources
cp = json.load(open(os.path.join(mbfc_path, "mbfc_conspiracy.json")))
fn = json.load(open(os.path.join(mbfc_path, "mbfc_fake-news.json")))
se = json.load(open(os.path.join(mbfc_path, "mbfc_satire.json")))
psc = json.load(open(os.path.join(mbfc_path, "mbfc_pseudoscience-dictionary.json")))

pseudo = [k[1] for k in psc]
len(pseudo)

conspiracy = [k[1] for k in cp]
len(conspiracy)

fakenews = [k[1] for k in fn]
len(fakenews)

satire = [k[1] for k in se]
len(satire)

notcr = list(set(conspiracy) | set(fakenews) | set(satire) | set(pseudo))
len(notcr)

unreliable = list(set(ns_fixed_) & set(notcr))
len(unreliable)

unreliable

# Political leaning

lf = json.load(open(os.path.join(mbfc_path, "mbfc_left.json")))
lc = json.load(open(os.path.join(mbfc_path, "mbfc_leftcenter.json")))

left = list(set(ns_fixed_) & (set([k[1] for k in lf]) | set([k[1] for k in lc])))
len(left)

rh = json.load(open(os.path.join(mbfc_path, "mbfc_right.json")))
rc = json.load(open(os.path.join(mbfc_path, "mbfc_right-center.json")))

right = (set(ns_fixed_) & (set([k[1] for k in rh]) | set([k[1] for k in rc])))
len(right)

right
