==> notebooks/Notebook14.Creating sets of labelled news sources.ipynb <==
# Creating sets of labelled news sources

I will create labelled news sources based on reliability, political leaning, media type, and geographical location.

import json
import csv
import os

import tldextract

from collections import Counter, defaultdict

## Reliability

Using News Guard

newsguard = json.load(open("new_newsguard_rated.json", "r", encoding="utf-8"))

reliability = dict()
for ns,rate in newsguard:
    if rate <= 60.0:
        reliability[ns] = "unreliable"
    else:
        reliability[ns] = "reliable"

json.dump(reliability, open("reliability.json", "w", encoding="utf-8"))

## Political Leaning

Using Media Bias Fact Check

pd1 = os.path.split(os.getcwd())[0]
mbfc_path = os.path.join(pd1, "data/mbfc")

lf = json.load(open(os.path.join(mbfc_path, "mbfc_left.json")))
lc = json.load(open(os.path.join(mbfc_path, "mbfc_leftcenter.json")))

rh = json.load(open(os.path.join(mbfc_path, "mbfc_right.json")))
rc = json.load(open(os.path.join(mbfc_path, "mbfc_right-center.json")))

ns_fixed_ = json.load(open("ns_fixed_.json", "r", encoding="utf-8"))

left = list(set(ns_fixed_) & (set([k[1] for k in lf]) | set([k[1] for k in lc])))
len(left)

right = (set(ns_fixed_) & (set([k[1] for k in rh]) | set([k[1] for k in rc])))
len(right)

political_leaning = dict()

for l in left:
    political_leaning[l] = "left"

for r in right:
    political_leaning[r] = "right"

json.dump(political_leaning, open("political_leaning.json", "w", encoding="utf-8"))

## Media Type

muckrack_path = os.path.join(pd1, "data/muckrack")

new_muckrack = json.load(open(os.path.join(muckrack_path, "all_new_muckrack.json")))
len(new_muckrack)

old_muckrack = json.load(open(os.path.join(muckrack_path, "all_old_muckrack.json")))

# let us clean both new and old muckrack
clean_new_muckrack = dict()

for nsm in new_muckrack:
    extracted = tldextract.extract(nsm)
    subdomain, domain, suffix = extracted
    # add both versions of domain.suffix and subdomain.domain.suffix
    full = ""
    # with subdomain
    if len(subdomain) > 0:
        #print(f"{subdomain}.{domain}.{suffix}")
        full = f"{subdomain}.{domain}.{suffix}"
        if len(full) > 0:
            clean_new_muckrack[full.replace("www.","").strip('/')] = new_muckrack[nsm]
    # without subdomain
    full = f"{domain}.{suffix}"
    if len(full) > 0:
        clean_new_muckrack[full.replace("www.","").strip('/')] = new_muckrack[nsm]
len(clean_new_muckrack)

clean_new_muckrack["tribunnews.com"]

len(set(clean_new_muckrack.keys()) & set(ns_fixed_))

len(ns_fixed_)

### According to new muck rack

media_type = dict()

for k in ns_fixed_:
    media_type[k] = clean_new_muckrack[k][0]['media_type'] if k in clean_new_muckrack else ''

media_type

len(set(media_type.values()))

Counter(media_type.values())

json.dump(media_type, open("media_type.json", "w", encoding="utf-8"))

### According to old muck rack


clean_old_muckrack = dict()

for nsm in old_muckrack:
    extracted = tldextract.extract(nsm)
    subdomain, domain, suffix = extracted
    # add both versions of domain.suffix and subdomain.domain.suffix
    full = ""
    # with subdomain
    if len(subdomain) > 0:
        #print(f"{subdomain}.{domain}.{suffix}")
        full = f"{subdomain}.{domain}.{suffix}"
        if len(full) > 0:
            clean_old_muckrack[full.replace("www.","").strip('/')] = old_muckrack[nsm]
    # without subdomain
    full = f"{domain}.{suffix}"
    if len(full) > 0:
        clean_old_muckrack[full.replace("www.","").strip('/')] = old_muckrack[nsm]
len(clean_old_muckrack)

media_type_current = dict()

for k in ns_fixed_:
    media_type_current[k] = clean_old_muckrack[k][0]['instanceof'] if k in clean_old_muckrack else ''

media_type_current

json.dump(media_type_current, open("media_type_current.json", "w", encoding="utf-8"))

## Country

country = dict()

for k in ns_fixed_:
    country[k] = clean_old_muckrack[k][0]['country'] if k in clean_old_muckrack else ''

Counter(country.values())

json.dump(country, open("country.json", "w", encoding="utf-8"))

clean_old_muckrack["reuters.com"]
